tiktok_partner_back/requirements.txt
appnope==0.1.4
asttokens==3.0.0
blinker==1.9.0
cached-property==2.0.1
cachetools==5.5.2
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.2.1
comm==0.2.1
cryptography==45.0.7
debugpy==1.8.16
decorator==5.2.1
defusedxml==0.7.1
dnspython==2.8.0
et_xmlfile==2.0.0
exceptiongroup==1.2.0
exchangelib==5.5.1
executing==2.2.1
Flask==3.1.2
Flask-Mail==0.10.0
google-api-core==2.25.1
google-api-python-client==2.181.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
googleapis-common-protos==1.70.0
greenlet==3.2.4
httplib2==0.31.0
idna==3.10
IMAPClient==3.0.1
ipykernel==6.30.1
ipython==8.30.0
isodate==0.7.2
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jupyter_client==8.6.3
jupyter_core==5.8.1
lxml==6.0.1
mail-parser==4.1.4
MarkupSafe==3.0.2
matplotlib-inline==0.1.6
nest-asyncio==1.6.0
numpy==2.2.6
oauthlib==3.3.1
openpyxl==3.1.5
packaging==25.0
pandas==2.3.2
parso==0.8.4
pexpect==4.9.0
pip==25.2
platformdirs==4.3.7
playwright==1.55.0
prompt-toolkit==3.0.43
proto-plus==1.26.1
protobuf==6.32.1
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pyee==13.0.0
Pygments==2.19.1
pyparsing==3.2.4
pyspnego==0.12.0
python-dateutil==2.9.0.post0
pytz==2025.2
pyzmq==26.2.0
requests==2.32.5
requests_ntlm==1.3.0
requests-oauthlib==2.0.0
rsa==4.9.1
setuptools==78.1.1
six==1.17.0
stack_data==0.6.3
tornado==6.5.1
traitlets==5.14.3
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.5.0
wcwidth==0.2.13
Werkzeug==3.1.3
wheel==0.45.1
flask==3.0.0
flask-cors==4.0.0
gunicorn==21.2.0

# FastAPI 相关依赖
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.18
sqlalchemy==2.0.36
pydantic-settings==2.6.1
email-validator==2.2.0
bcrypt==4.2.1